+++ demo/Tank.cpp
@@ -109,7 +109,7 @@ bool Tank::ProcessTouch( uint32 id, cpVect pos, TouchState state, bool rightClic
         }
         else if (state == ChipmunkDemo::TOUCH_MOVE)
         {
-            if (touchId == id)
+            if (touchId == (int)id)
             {
                 lastTouchPoint = pos;
                 return true;
@@ -117,7 +117,7 @@ bool Tank::ProcessTouch( uint32 id, cpVect pos, TouchState state, bool rightClic
         }
         else /* if (state == ChipmunkDemo::TOUCH_END) */
         {
-            if (touchId == id)
+            if (touchId == (int)id)
             {
                 touchId = -1;
                 lastTouchPoint = pos;
